# src/main/java/com/codev/api/security/auth/AuthRequest.java
package com.codev.api.security.auth;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;
import lombok.ToString;

@NoArgsConstructor
@AllArgsConstructor
@ToString
public class  AuthRequest {

    @Email
    public String email;

    @NotBlank
    public String password;

}